[dfs] fix and improve
Add ActionPlusAbstraction for DFSAction usage,

make exlo.build_action *not* wrap sentences that have semantic hints,
(so that DFS sentences aren't interfered with)

and make trie semantics run tests on @var's *if* there are constraints
on it. This is important for DFS actions, as to run a dfs with rule:

rule(::ρ):
 | $x |

 @x(::$y)?

 !! found.$y

end

needs this. if you were to do $x(::y)? instead,
it would *query from root*, which isn't what you want